Split protein panel into a new file
